requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-community
sentence-transformers==4.1.0
opensearch-py
pydantic
python-dotenv
openai